requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Neo4j
neo4j>=5.25.0

# Ollama (embedding)
ollama>=0.1.0

# MiniMax API
httpx>=0.26.0

# Document parsing
markdown>=3.5.0
python-frontmatter>=1.1.0

# Notion API
notion-client>=2.2.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
